requirements.txt
requests
pandas
python-dotenv
supabase
plotly

